backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "Phase II - FastAPI Todo Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlmodel>=0.0.22",
    "pyjwt>=2.9",
    "python-dotenv>=1.0",
    "bcrypt>=4.2",
]
